requirements.txt
fastapi==0.104.1
pydantic==2.5.0
uvicorn==0.24.0
mangum==0.17.0
boto3==1.34.10
python-dotenv==1.0.0